ref(outcomes): Hoist basic outcome types into module root
